move parts table to it's own component and start config of routes
